pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hqcb-hhh"
version = "0.1.0"
description = "HL-LHC (14 TeV) Higgs self-coupling (kappa_lambda) forecast toolkit with reproducible Asimov studies. Author: Oscar Fuentes Fernández."
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Oscar Fuentes Fernández"}]
license = {text = "AGPL-3.0-or-later"}
dependencies = ["numpy>=1.24", "scipy>=1.10", "pyyaml>=6.0", "matplotlib>=3.7"]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.6", "mypy>=1.8"]

# --- src/ layout: discovery ---
[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
strict = true